+++ Samples/SentrySampleShared/SentrySampleShared/RandomErrors.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+enum SampleError: Error {
+    case bestDeveloper
+    case happyCustomer
+    case awesomeCentaur
+}
+
+public class RandomErrorGenerator {
+    public static func generate() throws {
+        let random = Int.random(in: 0...2)
+        switch random {
+        case 0:
+            throw SampleError.bestDeveloper
+        case 1:
+            throw SampleError.happyCustomer
+        case 2:
+            throw SampleError.awesomeCentaur
+        default:
+            throw SampleError.bestDeveloper
+        }
+    }
+}