+++ Sources/Sentry/SentryHub.m
@@ -9,6 +9,7 @@
 #import "SentryFileManager.h"
 #import "SentryHub+Private.h"
 #import "SentryInstallation.h"
+#import "SentryIntegrationProtocol.h"
 #import "SentryLevelMapper.h"
 #import "SentryLog.h"
 #import "SentryNSTimerFactory.h"
@@ -792,7 +793,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)reportFullyDisplayed
 {
 #if SENTRY_HAS_UIKIT
-    [SentryUIViewControllerPerformanceTracker.shared reportFullyDisplayed];
+    [[SentryDependencyContainer.sharedInstance uiViewControllerPerformanceTracker]
+        reportFullyDisplayed];
 #endif // SENTRY_HAS_UIKIT
 }