+++ Sources/Sentry/SentrySessionReplayIntegration.m
@@ -104,9 +104,12 @@ static SentryTouchTracker *_touchTracker;
 
     id<SentryViewRenderer> viewRenderer;
     if (enableViewRendererV2) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Setting up view renderer v2, fast view rendering: %@",
+            enableFastViewRendering ? @"YES" : @"NO");
         viewRenderer =
             [[SentryViewRendererV2 alloc] initWithEnableFastViewRendering:enableFastViewRendering];
     } else {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Setting up default view renderer");
         viewRenderer = [[SentryDefaultViewRenderer alloc] init];
     }
     // We are using the flag for the experimental view renderer also for the experimental mask
@@ -117,6 +120,8 @@ static SentryTouchTracker *_touchTracker;
                                                     enableMaskRendererV2:enableViewRendererV2];
 
     if (touchTracker) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Setting up touch tracker, scale: %f", replayOptions.sizeScale);
         _touchTracker = [[SentryTouchTracker alloc] initWithDateProvider:_dateProvider
                                                                    scale:replayOptions.sizeScale];
         [self swizzleApplicationTouch];
@@ -163,6 +168,7 @@ static SentryTouchTracker *_touchTracker;
     NSData *lastReplay = [NSData dataWithContentsOfURL:lastReplayUrl];
 
     if (lastReplay == nil) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] No last replay info found");
         return nil;
     }
 
@@ -177,10 +183,13 @@ static SentryTouchTracker *_touchTracker;
  */
 - (void)resumePreviousSessionReplay:(SentryEvent *)event
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Resuming previous session replay");
     NSURL *dir = [self replayDirectory];
     NSDictionary<NSString *, id> *jsonObject = [self lastReplayInfo];
 
     if (jsonObject == nil) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] No last replay info found, not resuming previous session replay");
         return;
     }
 
@@ -200,8 +209,12 @@ static SentryTouchTracker *_touchTracker;
     int segmentId = hasCrashInfo ? crashInfo.segmentId + 1 : 0;
 
     if (type == SentryReplayTypeBuffer) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Previous session replay is a buffer, using error sample rate");
         float errorSampleRate = [jsonObject[@"errorSampleRate"] floatValue];
         if ([SentryDependencyContainer.sharedInstance.random nextNumber] >= errorSampleRate) {
+            SENTRY_LOG_INFO(
+                @"[Session Replay] Buffer session replay event not sampled, dropping replay");
             return;
         }
     }
@@ -218,6 +231,7 @@ static SentryTouchTracker *_touchTracker;
         ? [NSDate dateWithTimeIntervalSinceReferenceDate:crashInfo.lastSegmentEnd]
         : [resumeReplayMaker oldestFrameDate];
     if (beginning == nil) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] No frames to send, dropping replay");
         return; // no frames to send
     }
     NSDate *end = [beginning dateByAddingTimeInterval:duration];
@@ -254,8 +268,11 @@ static SentryTouchTracker *_touchTracker;
     NSError *_Nullable removeError;
     BOOL result = [NSFileManager.defaultManager removeItemAtURL:lastReplayURL error:&removeError];
     if (result == NO) {
-        SENTRY_LOG_ERROR(@"Can't delete '%@' with file item at url: '%@', reason: %@",
+        SENTRY_LOG_ERROR(
+            @"[Session Replay] Can't delete '%@' with file item at url: '%@', reason: %@",
             SENTRY_LAST_REPLAY, lastReplayURL, removeError);
+    } else {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Deleted last replay file at path: %@", lastReplayURL);
     }
 }
 
@@ -286,11 +303,14 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)startSession
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Starting session");
     [self.sessionReplay pause];
 
     _startedAsFullSession = [self shouldReplayFullSession:_replayOptions.sessionSampleRate];
 
     if (!_startedAsFullSession && _replayOptions.onErrorSampleRate == 0) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Not full session and onErrorSampleRate is 0, not starting session");
         return;
     }
 
@@ -300,9 +320,12 @@ static SentryTouchTracker *_touchTracker;
 - (void)runReplayForAvailableWindow
 {
     if (SentryDependencyContainer.sharedInstance.application.windows.count > 0) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Running replay for available window");
         // If a window its already available start replay right away
         [self startWithOptions:_replayOptions fullSession:_startedAsFullSession];
     } else if (@available(iOS 13.0, tvOS 13.0, *)) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Waiting for a scene to be available to started the replay");
         // Wait for a scene to be available to started the replay
         [_notificationCenter addObserver:self
                                 selector:@selector(newSceneActivate)
@@ -313,6 +336,7 @@ static SentryTouchTracker *_touchTracker;
 - (void)newSceneActivate
 {
     if (@available(iOS 13.0, tvOS 13.0, *)) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Scene is available, starting replay");
         [SentryDependencyContainer.sharedInstance.notificationCenterWrapper
             removeObserver:self
                       name:UISceneDidActivateNotification];
@@ -323,6 +347,7 @@ static SentryTouchTracker *_touchTracker;
 - (void)startWithOptions:(SentryReplayOptions *)replayOptions
              fullSession:(BOOL)shouldReplayFullSession
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Starting session");
     [self startWithOptions:replayOptions
          screenshotProvider:_currentScreenshotProvider ?: _viewPhotographer
         breadcrumbConverter:_currentBreadcrumbConverter
@@ -335,11 +360,13 @@ static SentryTouchTracker *_touchTracker;
      breadcrumbConverter:(id<SentryReplayBreadcrumbConverter>)breadcrumbConverter
              fullSession:(BOOL)shouldReplayFullSession
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Starting session");
     NSURL *docs = [self replayDirectory];
     NSString *currentSession = [NSUUID UUID].UUIDString;
     docs = [docs URLByAppendingPathComponent:currentSession];
 
     if (![NSFileManager.defaultManager fileExistsAtPath:docs.path]) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] Creating directory at path: %@", docs.path);
         [NSFileManager.defaultManager createDirectoryAtURL:docs
                                withIntermediateDirectories:YES
                                                 attributes:nil
@@ -398,6 +425,8 @@ static SentryTouchTracker *_touchTracker;
                           path:(NSString *)path
                        options:(SentryReplayOptions *)options
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Saving current session info for session: %@ to path: %@",
+        sessionId, path);
     NSDictionary *info =
         [[NSDictionary alloc] initWithObjectsAndKeys:sessionId.sentryIdString, @"replayId",
             path.lastPathComponent, @"path", @(options.onErrorSampleRate), @"errorSampleRate", nil];
@@ -407,16 +436,20 @@ static SentryTouchTracker *_touchTracker;
     NSString *infoPath = [[path stringByDeletingLastPathComponent]
         stringByAppendingPathComponent:SENTRY_CURRENT_REPLAY];
     if ([NSFileManager.defaultManager fileExistsAtPath:infoPath]) {
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Removing existing current replay info at path: %@", infoPath);
         [NSFileManager.defaultManager removeItemAtPath:infoPath error:nil];
     }
     [data writeToFile:infoPath atomically:YES];
 
+    SENTRY_LOG_DEBUG(@"[Session Replay] Saved current session info at path: %@", infoPath);
     sentrySessionReplaySync_start([[path stringByAppendingPathComponent:@"crashInfo"]
         cStringUsingEncoding:NSUTF8StringEncoding]);
 }
 
 - (void)moveCurrentReplay
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Moving current replay");
     NSFileManager *fileManager = NSFileManager.defaultManager;
 
     NSURL *path = [self replayDirectory];
@@ -425,30 +458,33 @@ static SentryTouchTracker *_touchTracker;
 
     NSError *error;
     if ([fileManager fileExistsAtPath:last.path]) {
-        SENTRY_LOG_DEBUG(@"Removing last replay file at path: %@", last);
+        SENTRY_LOG_DEBUG(@"[Session Replay] Removing last replay file at path: %@", last);
         if ([NSFileManager.defaultManager removeItemAtURL:last error:&error] == NO) {
-            SENTRY_LOG_ERROR(@"Could not delete last replay file, reason: %@", error);
+            SENTRY_LOG_ERROR(
+                @"[Session Replay] Could not delete last replay file, reason: %@", error);
             return;
         }
-        SENTRY_LOG_DEBUG(@"Removed last replay file at path: %@", last);
+        SENTRY_LOG_DEBUG(@"[Session Replay] Removed last replay file at path: %@", last);
     } else {
-        SENTRY_LOG_DEBUG(@"No last replay file to remove at path: %@", last);
+        SENTRY_LOG_DEBUG(@"[Session Replay] No last replay file to remove at path: %@", last);
     }
 
     if ([fileManager fileExistsAtPath:current.path]) {
-        SENTRY_LOG_DEBUG(@"Moving current replay file at path: %@ to: %@", current, last);
+        SENTRY_LOG_DEBUG(
+            @"[Session Replay] Moving current replay file at path: %@ to: %@", current, last);
         if ([fileManager moveItemAtURL:current toURL:last error:&error] == NO) {
-            SENTRY_LOG_ERROR(@"Could not move replay file, reason: %@", error);
+            SENTRY_LOG_ERROR(@"[Session Replay] Could not move replay file, reason: %@", error);
             return;
         }
-        SENTRY_LOG_DEBUG(@"Moved current replay file at path: %@", current);
+        SENTRY_LOG_DEBUG(@"[Session Replay] Moved current replay file at path: %@", current);
     } else {
-        SENTRY_LOG_DEBUG(@"No current replay file to move at path: %@", current);
+        SENTRY_LOG_DEBUG(@"[Session Replay] No current replay file to move at path: %@", current);
     }
 }
 
 - (void)cleanUp
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Cleaning up");
     NSURL *replayDir = [self replayDirectory];
     NSDictionary<NSString *, id> *lastReplayInfo = [self lastReplayInfo];
     NSString *lastReplayFolder = lastReplayInfo[@"path"];
@@ -458,6 +494,7 @@ static SentryTouchTracker *_touchTracker;
     // listing files and creating a new replay session.
     NSArray *replayFiles = [fileManager allFilesInFolder:replayDir.path];
     if (replayFiles.count == 0) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] No replay files to clean up");
         return;
     }
 
@@ -465,6 +502,7 @@ static SentryTouchTracker *_touchTracker;
         for (NSString *file in replayFiles) {
             // Skip the last replay folder.
             if ([file isEqualToString:lastReplayFolder]) {
+                SENTRY_LOG_DEBUG(@"[Session Replay] Skipping last replay folder: %@", file);
                 continue;
             }
 
@@ -472,6 +510,8 @@ static SentryTouchTracker *_touchTracker;
 
             // Check if the file is a directory before deleting it.
             if ([fileManager isDirectory:filePath]) {
+                SENTRY_LOG_DEBUG(
+                    @"[Session Replay] Removing replay directory at path: %@", filePath);
                 [fileManager removeFileAtPath:filePath];
             }
         }
@@ -480,24 +520,28 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)pause
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Pausing session");
     [self.sessionReplay pause];
 }
 
 - (void)resume
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Resuming session");
     [self.sessionReplay resume];
 }
 
 - (void)start
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Starting session");
     if (_rateLimited) {
-        SENTRY_LOG_WARN(
-            @"This session was rate limited. Not starting session replay until next app session");
+        SENTRY_LOG_WARN(@"[Session Replay] This session was rate limited. Not starting session "
+                        @"replay until next app session");
         return;
     }
 
     if (self.sessionReplay != nil) {
         if (self.sessionReplay.isFullSession == NO) {
+            SENTRY_LOG_DEBUG(@"[Session Replay] Not full session, capturing replay");
             [self.sessionReplay captureReplay];
         }
         return;
@@ -509,12 +553,14 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)stop
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Stopping session");
     [self.sessionReplay pause];
     self.sessionReplay = nil;
 }
 
 - (void)sentrySessionEnded:(SentrySession *)session
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Session ended");
     [self pause];
     [_notificationCenter removeObserver:self
                                    name:UIApplicationDidEnterBackgroundNotification
@@ -527,18 +573,21 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)sentrySessionStarted:(SentrySession *)session
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Session started");
     _rateLimited = NO;
     [self startSession];
 }
 
 - (BOOL)captureReplay
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Capturing replay");
     return [self.sessionReplay captureReplay];
 }
 
 - (void)configureReplayWith:(nullable id<SentryReplayBreadcrumbConverter>)breadcrumbConverter
          screenshotProvider:(nullable id<SentryViewScreenshotProvider>)screenshotProvider
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Configuring replay");
     if (breadcrumbConverter) {
         _currentBreadcrumbConverter = breadcrumbConverter;
         self.sessionReplay.breadcrumbConverter = breadcrumbConverter;
@@ -552,6 +601,7 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)setReplayTags:(NSDictionary<NSString *, id> *)tags
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Setting replay tags: %@", tags);
     self.sessionReplay.replayTags = [tags copy];
 }
 
@@ -562,6 +612,7 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)uninstall
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Uninstalling");
     [SentrySDK.currentHub unregisterSessionListener:self];
     _touchTracker = nil;
     [self pause];
@@ -579,6 +630,7 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)swizzleApplicationTouch
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Swizzling application touch tracker");
 #    pragma clang diagnostic push
 #    pragma clang diagnostic ignored "-Wshadow"
     SEL selector = NSSelectorFromString(@"sendEvent:");
@@ -604,6 +656,9 @@ static SentryTouchTracker *_touchTracker;
                                                 level:(enum SentryLevel)level
                                                  data:(nullable NSDictionary<NSString *, id> *)data
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Creating breadcrumb with timestamp: %@, category: %@, "
+                     @"message: %@, level: %lu, data: %@",
+        timestamp, category, message, level, data);
     return [[SentryRRWebBreadcrumbEvent alloc] initWithTimestamp:timestamp
                                                         category:category
                                                          message:message
@@ -617,6 +672,9 @@ static SentryTouchTracker *_touchTracker;
                                                  description:(NSString *)description
                                                         data:(NSDictionary<NSString *, id> *)data
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Creating network breadcrumb with timestamp: %@, "
+                     @"endTimestamp: %@, operation: %@, description: %@, data: %@",
+        timestamp, endTimestamp, operation, description, data);
     return [[SentryRRWebSpanEvent alloc] initWithTimestamp:timestamp
                                               endTimestamp:endTimestamp
                                                  operation:operation
@@ -626,6 +684,7 @@ static SentryTouchTracker *_touchTracker;
 
 + (id<SentryReplayBreadcrumbConverter>)createDefaultBreadcrumbConverter
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Creating default breadcrumb converter");
     return [[SentrySRDefaultBreadcrumbConverter alloc] init];
 }
 
@@ -641,6 +700,8 @@ static SentryTouchTracker *_touchTracker;
                                replayRecording:(SentryReplayRecording *)replayRecording
                                       videoUrl:(NSURL *)videoUrl
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] New segment with replay event, eventId: %@, segmentId: %lu",
+        replayEvent.eventId, replayEvent.segmentId);
     if ([_rateLimits isRateLimitActive:kSentryDataCategoryReplay] ||
         [_rateLimits isRateLimitActive:kSentryDataCategoryAll]) {
         SENTRY_LOG_DEBUG(
@@ -660,6 +721,7 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)sessionReplayStartedWithReplayId:(SentryId *)replayId
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Session replay started with replay id: %@", replayId);
     [SentrySDK.currentHub configureScope:^(
         SentryScope *_Nonnull scope) { scope.replayId = [replayId sentryIdString]; }];
 }
@@ -681,13 +743,15 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)showMaskPreview:(CGFloat)opacity
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Showing mask preview with opacity: %f", opacity);
     if ([SentryDependencyContainer.sharedInstance.crashWrapper isBeingTraced] == NO) {
+        SENTRY_LOG_DEBUG(@"[Session Replay] No tracing is active, not showing mask preview");
         return;
     }
 
     UIWindow *window = SentryDependencyContainer.sharedInstance.application.windows.firstObject;
     if (window == nil) {
-        SENTRY_LOG_WARN(@"There is no UIWindow available to display preview");
+        SENTRY_LOG_WARN(@"[Session Replay] No UIWindow available to display preview");
         return;
     }
 
@@ -702,6 +766,7 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)hideMaskPreview
 {
+    SENTRY_LOG_DEBUG(@"[Session Replay] Hiding mask preview");
     [_previewView removeFromSuperview];
     _previewView = nil;
 }
@@ -710,7 +775,8 @@ static SentryTouchTracker *_touchTracker;
 
 - (void)connectivityChanged:(BOOL)connected typeDescription:(nonnull NSString *)typeDescription
 {
-
+    SENTRY_LOG_DEBUG(@"[Session Replay] Connectivity changed to: %@, type: %@",
+        connected ? @"connected" : @"disconnected", typeDescription);
     if (connected) {
         [_sessionReplay resume];
     } else {