+++ Sources/Sentry/include/HybridPublic/SentryBaseIntegration.h
@@ -1,5 +1,11 @@
 #import <Foundation/Foundation.h>
 
+#if __has_include(<Sentry/SentryIntegrationProtocol.h>)
+#    import <Sentry/SentryIntegrationProtocol.h>
+#else
+#    import "SentryIntegrationProtocol.h"
+#endif
+
 NS_ASSUME_NONNULL_BEGIN
 
 typedef NS_OPTIONS(NSUInteger, SentryIntegrationOption) {
@@ -23,11 +29,12 @@ typedef NS_OPTIONS(NSUInteger, SentryIntegrationOption) {
     kIntegrationOptionEnableCrashHandler = 1 << 16,
     kIntegrationOptionEnableMetricKit = 1 << 17,
     kIntegrationOptionEnableReplay = 1 << 18,
+    kIntegrationOptionStartFramesTracker = 1 << 19,
 };
 
 @class SentryOptions;
 
-@interface SentryBaseIntegration : NSObject
+@interface SentryBaseIntegration : NSObject <SentryIntegrationProtocol>
 
 - (NSString *)integrationName;
 - (BOOL)installWithOptions:(SentryOptions *)options;