+++ Sources/Swift/Integrations/SessionReplay/SentryReplayType.swift
@@ -22,3 +22,9 @@ extension SentryReplayType {
         }
     }
 }
+
+extension SentryReplayType: CustomStringConvertible {
+    var description: String {
+        return toString()
+    }
+}