+++ Utils/VersionBump/main.swift
@@ -16,9 +16,10 @@ let files = [
 // Files that only accept the format x.x.x in order to release an app using the framework.
 // This will enable publishing apps with SDK beta version.
 let restrictFiles = [
-    "./Samples/iOS-Swift/iOS-Swift/Sample.xcconfig",
     "./Sources/Configuration/SDK.xcconfig",
-    "./Sources/Configuration/SentrySwiftUI.xcconfig"
+    "./Sources/Configuration/Versioning.xcconfig", 
+    "./Sources/Configuration/SentrySwiftUI.xcconfig",
+    "./Samples/Shared/Config/Versioning.xcconfig"
 ]
 
 let args = CommandLine.arguments